move get_timerange_from_exdict
